# Speed up Circuit.s_external property(method) by using block matrix ca…