@jhillairet 2024-05-13
Thank you for this PR. 

I'm quite puzzled: I have hard time to understand why your proposal is faster than a extracting a subarray from an array... I think it adds a lot of obfuscation to what was a simple and understable implementation... 

Actually, isn't `meshgrid` taking most of the time?