@Asachoo 2024-05-13
> Thank you for this PR.
> 
> I'm quite puzzled: I have hard time to understand why your proposal is faster than a extracting a subarray from an array... I think it adds a lot of obfuscation to what was a simple and understable implementation...
> 
> Edit: OK, I see now. You're skipping the `.s` method. Interesting approach.

Apologies for the confusion caused by the ambiguous variable names and function intentions in the modified code. I overlooked the maintenance of variable names and comments before submitting the PR.

As we are only interested in the final computed S-parameters of the external ports, I hypothesized that speeding up s_external could be achieved through [block matrix manipulation](https://malithjayaweera.com/2020/07/blocked-matrix-multiplication/) . The intention of this PR was to treat self.X and self.C as two  block matrices based on external ports and inner ports (although my poor coding execution undermined this idea).

![](https://malithjayaweera.com/wp-content/uploads/2020/07/matrix-matrix-blocked.jpg)